// contester/ClientApp/src/generated/client/model/contestParticipationDto.ts
/**
 * contester
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ContestParticipationDto { 
    id?: string;
    name?: string;
    description?: string;
    isPublic?: boolean;
    createdAt?: Date;
    startDate?: Date;
    finishDate?: Date;
    readonly timeUntilStartSeconds?: number;
    readonly timeUntilFinishSeconds?: number;
    authorId?: string;
    userParticipates?: boolean;
}